Moves Msg::un_preparsecode() to Parser::getEditableString()
